django/requirements.txt
wagtail==1.9
wagtailmenus==2.2.1
django-extensions==1.7.7
social-auth-app-django==1.1.0
django-rest-auth==0.9.1
djangorestframework-jsonapi==2.1.1
djangorestframework-jwt==1.9.0
django-rest-swagger==2.1.2
django-reversion==2.0.8
psycopg2==2.7.1
pygit2==0.24.2
django-location-field==2.0.2
django-model-utils==2.6.1
django-timezone-field==2.0
bagit==1.5.4
semver==2.7.6
django-guardian==1.4.6
django-jinja==2.2.2
django-registration==2.2
invoke
django-webpack-loader
rarfile
# need to skip dependency checking, scandir fails pip3 install in the container
# hashfs==0.7.0
# CoMSES citation app
-e git+https://github.com/comses/citation.git#egg=citation
# NLP Summary dependencies
sumy==0.6.0
bleach==2.0.0
markdown==2.6.8
